Refactor ScatterOptions into its own source file.
The SeriesOptions source file was getting too crowded.
